Add support for writing metrics to InfluxDB v2
✨ Start both v1 and v2 Docker containers

Start separate Docker containers for both Influx v1 and v2.

The v2 Docker containers aren't yet being published to DockerHub but are available on quay.io, so we're pulling from there for now.

While v2 is in beta, it uses port 9999 by default, so that's what we're using here.  Once v2 is released, that will change and we'll have to adapt to that in a future PR.

The v2 Docker container does not currently auto-provision itself using environment variables, so we're running an explicit curl command to do that.

All of the curl commands write to separate error log files so that no potentially-helpful debugging information gets lost.

Also, added the error-*.log files to `.gitignore` to keep them from showing up as changed.

Co-authored-by: Iris Scholten <iris@influxdata.com>

✏️ Fix minor typos

Co-authored-by: Iris Scholten <iris@influxdata.com>

🚧 Annotate code with TODOs for Influx v2

Co-authored-by: Iris Scholten <iris@influxdata.com>

✅ Add tests for v2 option validation

Co-authored-by: Iris Scholten <iris@influxdata.com>

📝 Update moduledocs and type for new options

✨ Update options to include a version

This option allows the version of influxdb to be explicitly specified

✨ Ensure token is provided for v2 options

This is required for making requests to InfluxDB v2 to ensure authentication

✨ Save parsed token from setup command in .token file

This token is generated every time a user is provisioned.
This is only useful for tests to make sure we can make authenticated requests to the running InfluxDB v2.

✨ Use token from .token file in test suite

Co-authored-by: Delmer Reed <delmer814+1@gmail.com>

📝 Update moduledoc with token info

Co-authored-by: Delmer Reed <delmer814+1@gmail.com>

✨ Add default and validation for version

Co-authored-by: Delmer Reed <delmer814+1@gmail.com>

♻️  Pass options into start reporter

Co-authored-by: Delmer Reed <delmer814+1@gmail.com>

♻️  update make_options to take version and protocol

♻️  v1/v2 client and refute_reported

wip

♻️  refactor to use one default config

Refactors the tests to use one default config and updates
refute_reported to use pattern matching on config version

Co-authored-by: Randy Coulman <rcoulman@influxdata.com>

♻️ Use separate config for assertions/cleanup

We do all asserting and cleaning up using HTTP, so we need a HTTP-based config for that.

We now pass the test context to various functions and allow them to make the appropriate config based on that.

Co-authored-by: Delmer Reed <delmer814+1@gmail.com>

♻️ Re-order helper functions in test

🚧 WIP: build up assertion functions

✨ Implement cleanup for v2 API

Includes refactoring test helper functions to make them easier to use and remove duplication.

Co-authored-by: Iris Scholten <iris@influxdata.com>

✨ Add V2 event handler code

Still building out assertion infrastructure for v2.  Incorporated a flux table parser that will be extracted to its own public repo and published to hex.pm before we submit this PR.

Co-authored-by: Iris Scholten <iris@influxdata.com>

✨: Implement assert_reported for v2 flux responses

Flux parser was updated to separate out every table into its own map

♻️ Refactor flux parsing to be simpler and more specific to this use

🐛 Fix DateTime updating for older elixir versiond

✏️ Fix typo

📝 Update moduledoc description of db for clarity

✨ Update error message to specify v1

♻️ Update the order of validate_v2_params! to have happy path first

📝 Update read me to include version options

♻️ Update parameter order for flux parser

📝 Update readme and options docs

🐛 Try using DateTime add again

️ Use NaiveDateTime in order to support older elixir versions

Co-authored-by: Iris Scholten <iris@influxdata.com>
Co-authored-by: Randy Coulman <rcoulman@influxdata.com>
Co-authored-by: Delmer Reed <delmer814+1@gmail.com>